Removal of coupling to app, additional tests
This should be the last commit before this branch can be committed. This
removes the coupling to the lamia app, making the email code stand
alone. This is to allow for easier building of the test enviroment.
Two new dependancies were added so that asyncronous email send testing
could be performed properly.